requirements.txt
boto3==1.7.80
botocore==1.10.80
chardet==2.2.1
colorama==0.2.5
confusable-homoglyphs==3.1.1
dj-database-url==0.5.0
Django==1.11
django-forms-bootstrap==3.1.0
django-registration==2.4.1
django-storages==1.6.6
docutils==0.14
easy-thumbnails==2.5
gunicorn==19.9.0
html5lib==0.999
jmespath==0.9.3
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pycurl==7.19.3
python-dateutil==2.7.3
pytz==2018.5
requests==2.2.1
s3transfer==0.1.13
six==1.5.2
stripe==2.5.0
urllib3==1.7.1
whitenoise==4.0
